refactor: extract Harness interface, wrap Docs logic as DocsHarness (#97)
